// tests/test-postgres/src/codegen/vexnor_dev.account-table.ts
/*
 File generated by Vexnor. Do not edit! 
*/
import * as vexnor from "@vexnor/core";
import * as udt from "./vexnor_dev-enums.js";

export const Account = vexnor.newSqlTable<{
   Select: IAccountSelect;
   Insert: IAccountInsert;
   Update: IAccountUpdate;
   Delete: true;
}>( {
   crud: {
      select: true,
      insert: true,
      update: true,
      delete: true,
   },
   tableInfo: {
      name: "account",
      schema: "vexnor_dev",
   },
   pk: ["accountId"],
   dialect: "postgresql",
   source: "@vexnor/test-postgres:src/codegen",
   columns: {

      /**
       * account_id uuid default gen_random_uuid()
       */
      accountId: "account_id",

      /**
       * status account_status default 'created'::vexnor_dev.account_status
       */
      status: "status",

      /**
       * email varchar
       */
      email: "email",

      /**
       * first_name varchar
       */
      firstName: "first_name",

      /**
       * last_name varchar
       */
      lastName: "last_name",

      /**
       * notes text
       */
      notes: "notes",

      /**
       * created_at timestamptz default now()
       */
      createdAt: "created_at",

      /**
       * modified_at timestamptz default now()
       */
      modifiedAt: "modified_at",

      /**
       * parent_id uuid
       */
      parentId: "parent_id",
   },
   jsonSchema: {
      createdAt: "Date",
      modifiedAt: "Date",
   },
   fk: [
      { from: ["parentId"], to: { schema: "vexnor_dev", table: "account", columns: ["accountId"] } },
   ],
   dbSchema: {
      accountId: { dbType: "uuid", type: vexnor.SqlLiteralType.String, default: "gen_random_uuid()" },
      status: { dbType: "account_status", type: vexnor.SqlLiteralType.Udt, default: "'created'::vexnor_dev.account_status", values: ["created", "confirmed", "deleted"] },
      email: { dbType: "varchar", type: vexnor.SqlLiteralType.String },
      firstName: { dbType: "varchar", type: vexnor.SqlLiteralType.String },
      lastName: { dbType: "varchar", type: vexnor.SqlLiteralType.String },
      notes: { dbType: "text", type: vexnor.SqlLiteralType.String, nullable: true },
      createdAt: { dbType: "timestamptz", type: vexnor.SqlLiteralType.Date, default: "now()" },
      modifiedAt: { dbType: "timestamptz", type: vexnor.SqlLiteralType.Date, default: "now()" },
      parentId: { dbType: "uuid", type: vexnor.SqlLiteralType.String, nullable: true },
   },
});
export type IAccountInsert = {
   accountId?: string;
   status?: udt.AccountStatusUdt;
   email: string;
   firstName: string;
   lastName: string;
   notes?: string | null;
   createdAt?: Date;
   modifiedAt?: Date;
   parentId?: string | null;
};

export type IAccountUpdate = Partial<IAccountInsert>;

export type IAccountSelect = {
   accountId: string;
   status: udt.AccountStatusUdt;
   email: string;
   firstName: string;
   lastName: string;
   notes: string | null;
   createdAt: Date;
   modifiedAt: Date;
   parentId: string | null;
};

export type IAccountJson = vexnor.JsonRow<IAccountSelect>;