// tests/test-postgres/src/codegen/vexnor_dev.product-table.ts
/*
 File generated by Vexnor. Do not edit! 
*/
import * as vexnor from "@vexnor/core";


export const Product = vexnor.newSqlTable<{
   Select: IProductSelect;
   Insert: IProductInsert;
   Update: IProductUpdate;
   Delete: true;
}>( {
   crud: {
      select: true,
      insert: true,
      update: true,
      delete: true,
   },
   tableInfo: {
      name: "product",
      schema: "vexnor_dev",
   },
   pk: ["productId"],
   dialect: "postgresql",
   source: "@vexnor/test-postgres:src/codegen",
   columns: {

      /**
       * product_id uuid default gen_random_uuid()
       */
      productId: "product_id",

      /**
       * created_at timestamptz default now()
       */
      createdAt: "created_at",

      /**
       * modified_at timestamptz default now()
       */
      modifiedAt: "modified_at",

      /**
       * label varchar
       */
      label: "label",

      /**
       * price numeric
       */
      price: "price",

      /**
       * discount numeric
       */
      discount: "discount",

      /**
       * is_available bool default true
       */
      isAvailable: "is_available",

      /**
       * is_published bool default false
       */
      isPublished: "is_published",

      /**
       * metadata jsonb
       */
      metadata: "metadata",

      /**
       * tags _text
       */
      tags: "tags",
   },
   jsonSchema: {
      createdAt: "Date",
      modifiedAt: "Date",
   },
   dbSchema: {
      productId: { dbType: "uuid", type: vexnor.SqlLiteralType.String, default: "gen_random_uuid()" },
      createdAt: { dbType: "timestamptz", type: vexnor.SqlLiteralType.Date, default: "now()" },
      modifiedAt: { dbType: "timestamptz", type: vexnor.SqlLiteralType.Date, default: "now()" },
      label: { dbType: "varchar", type: vexnor.SqlLiteralType.String },
      price: { dbType: "numeric", type: vexnor.SqlLiteralType.String },
      discount: { dbType: "numeric", type: vexnor.SqlLiteralType.String, nullable: true },
      isAvailable: { dbType: "bool", type: vexnor.SqlLiteralType.Boolean, default: "true" },
      isPublished: { dbType: "bool", type: vexnor.SqlLiteralType.Boolean, default: "false" },
      metadata: { dbType: "jsonb", type: vexnor.SqlLiteralType.Json, nullable: true },
      tags: { dbType: "_text", type: vexnor.SqlLiteralType.Unknown, nullable: true },
   },
});
export type IProductInsert = {
   productId?: string;
   createdAt?: Date;
   modifiedAt?: Date;
   label: string;
   price: string;
   discount?: string | null;
   isAvailable?: boolean;
   isPublished?: boolean;
   metadata?: unknown | null;
   tags?: unknown | null;
};

export type IProductUpdate = Partial<IProductInsert>;

export type IProductSelect = {
   productId: string;
   createdAt: Date;
   modifiedAt: Date;
   label: string;
   price: string;
   discount: string | null;
   isAvailable: boolean;
   isPublished: boolean;
   metadata: unknown | null;
   tags: unknown | null;
};

export type IProductJson = vexnor.JsonRow<IProductSelect>;